backend/requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
python-telegram-bot==21.1.1
firebase-admin==6.5.0
requests==2.31.0
beautifulsoup4==4.14.3
